requirements.txt
dict2xml==1.7.0
fastavro==1.3.4
lxml==4.6.3
PyYAML==5.4.1
simplejson==3.17.2
tabulate==0.8.9
